+++ app/src/main/java/com/wafflestudio/snutt2/components/compose/SegmentedPicker.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.unit.sp
 import com.wafflestudio.snutt2.R
 import com.wafflestudio.snutt2.ui.SNUTTColors
 import com.wafflestudio.snutt2.ui.SNUTTTypography
-import com.wafflestudio.snutt2.views.logged_in.home.settings.LectureReminderOffset
+import com.wafflestudio.snutt2.domainmodel.LectureReminderOffset
 
 /**
  * 여러 선택지 중 하나를 선택할 수 있는 UI 컴포넌트
