+++ app/src/main/java/com/wafflestudio/snutt2/domainmodel/LectureWithReminderOption.kt
@@ -1,7 +1,5 @@
 package com.wafflestudio.snutt2.domainmodel
 
-import com.wafflestudio.snutt2.views.logged_in.home.settings.LectureReminderOffset
-
 data class LectureWithReminderOption(
     val lectureId: String,
     val lectureTitle: String,
